apps/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
rembg==2.0.50
replicate==0.22.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
